test compression e2e
